Add 'seed-installer' Docker container entrypoint.sh script
The `entrypoint` for the 'seed-installer' Docker container is a local
file reference via a Bind mount to this `entrypoint.sh` script. A Bind
mount was favoured over "baking" in the `entrypoint.sh` script into the
container image file for the ability to alter it outside without a
Docker image rebuild.

A local Python3 webserver is ran in order to satisfy `debian-cd`'s
`DI_WWW_HOME` environment variable responsible for downloading the core
Debian Installer (d-i) components.