[debian-cd] Dynamically specify Debian release version
The `entrypoint.sh` script determines the Debian release version as part
of its `prepenv()` function. It does this by examining the corresponding
value of the 'Version:' entry in the local Debian mirror's `Release`
file and exporting the result as an environment variable:
`DEBIAN_RELEASE_VERSION`.

This functionality was added to remove the manual overhead of specifying
the Debian release version in CONF.sh prior to building the customised
Debian installer ISO.