[preseed] Set networking answers
The targeted "seed" host requires internet connectivity in order to
synchronise time (NTP) and clone the `grindirion/puppet-bootstrap` Git
repository.

Static IPv4 configuration is explicitly specified to avoid depending on
an existing DHCP environment; the "seed" host is assumed to be the first
host provisioned in an environment. A '/16' subnet was chosen as it
offered a sufficiently large pool for host assignment.

The "seed" host will operate as the initial nameserver to provide the
DNS service to all subsequently deployed hosts.

As per the original `seed-installer` design considerations *no* external
mirror is specified during "seed" host installation.